api/requirements.txt
Flask==2.2.2
SQLAlchemy==1.3.23
PyMySQL==1.0.2
Faker==8.0.0
PyJWT==2.3.0
cryptography==37.0.2

autoflake==1.3
black==22.3.0
isort==5.0.3
pylint==2.6.0
pytest==4.6.3
pytest-cov==2.6.1
Flask-Cors == 3.0.10